+++ buybackprogram/__init__.py
@@ -2,5 +2,5 @@
 
 default_app_config = "buybackprogram.apps.BuybackProgramConfig"
 
-__version__ = "1.12.3"
+__version__ = "1.13.0"
 __title__ = "Buyback Program"